% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{plot.mlp_importance}
\alias{plot.mlp_importance}
\title{Plot permutation importance}
\usage{
\method{plot}{mlp_importance}(x, ...)
}
\arguments{
\item{x}{An \code{mlp_importance} object.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot permutation importance
}
