% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mlp_fit}
\alias{predict.mlp_fit}
\title{Predict from a fitted MLP}
\usage{
\method{predict}{mlp_fit}(object, new_data, type = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{mlp_fit} object.}

\item{new_data}{New predictor data.}

\item{type}{Prediction type.}

\item{...}{Unused.}
}
\value{
Predictions in a task-appropriate format.
}
\description{
Predict from a fitted MLP
}
